divided the code up into more files. added css to map
