+++ src/popup/views/popup.js
@@ -1,59 +1,14 @@
-/* eslint-disable jsx-a11y/click-events-have-key-events */
 import React, { Component } from 'react';
-import {
-    Col, Grid, Row, ListGroup,
-} from 'react-bootstrap';
-import uuid from 'uuid';
+import { Col, Grid, Row } from 'react-bootstrap';
 import PopupMap from './popupMap';
 import PopupNavbar from './navbar';
-import { removeWhere } from '../../utils';
-import './resultsList.css';
+import ResultsList from './resultsList';
 
-function ListItem({ children, onClick }) {
-    return (
-        // eslint-disable-next-line jsx-a11y/no-noninteractive-element-interactions
-        <li
-            className="list-group-item list-item"
-            onClick={(ev) => {
-                if (ev.target.tagName === 'LI') {
-                    onClick();
-                }
-            }}
-        >
-            {children}
-        </li>
-    );
-}
-
-const PopupMapBox = ({ placesScraped }) => (
-    <div style={{ height: '400px' }}>
-        {placesScraped.length === 0 ? (
-            <div className="map-absent-box" />
-        ) : (
-            <PopupMap placesScraped={placesScraped} />
-        )}
-    </div>
-);
-
-const ResultsList = ({ places, setPlaces, setSelectedPlace }) => (
-    <div className="results-list">
-        <ListGroup>
-            {places.map(place => (
-                <ListItem key={uuid.v4()} onClick={() => setSelectedPlace(place)}>
-                    <button
-                        type="button"
-                        className="icon-button"
-                        onClick={() => setPlaces(removeWhere(places, 'name', place.name))}
-                    >
-                        <span className="glyphicon glyphicon-remove" />
-                    </button>
-                    {' '}
-                    {place.name}
-                </ListItem>
-            ))}
-        </ListGroup>
-    </div>
-);
+// const PopupMapBox = ({ placesScraped }) => (
+//     <div style={{ height: '400px' }}>
+//         {placesScraped.length !== 0 && <PopupMap placesScraped={placesScraped} />}
+//     </div>
+// );
 
 class Popup extends Component {
     constructor(props) {
@@ -85,31 +40,33 @@ class Popup extends Component {
 
     render() {
         return (
-            <Grid>
-                <Row>
-                    <PopupNavbar
-                        setPlaces={this.setLastScrapedPlaces}
-                        selectedPlace={this.state.selectedPlace}
-                        setLastPlacesScraped={this.setLastPlacesScraped}
-                    />
-                </Row>
-                {this.state.placesScraped.length !== 0 && (
-                    <>
-                        <Row>
-                            <Col xs={8}>
-                                <PopupMapBox placesScraped={this.state.placesScraped} />
-                            </Col>
-                            <Col xs={4}>
-                                <ResultsList
-                                    places={this.state.placesScraped}
-                                    setPlaces={this.setLastPlacesScraped}
-                                    setSelectedPlace={this.setSelectedPlace}
-                                />
-                            </Col>
-                        </Row>
-                    </>
-                )}
-            </Grid>
+            <>
+                <PopupNavbar
+                    setPlaces={this.setLastScrapedPlaces}
+                    selectedPlace={this.state.selectedPlace}
+                    setLastPlacesScraped={this.setLastPlacesScraped}
+                />
+                <Grid>
+                    {this.state.placesScraped.length !== 0 && (
+                        <>
+                            <Row>
+                                <Col xs={7}>
+                                    {this.state.placesScraped.length !== 0 && (
+                                        <PopupMap placesScraped={this.state.placesScraped} />
+                                    )}
+                                </Col>
+                                <Col xs={4} xsOffset={1}>
+                                    <ResultsList
+                                        places={this.state.placesScraped}
+                                        setPlaces={this.setLastPlacesScraped}
+                                        setSelectedPlace={this.setSelectedPlace}
+                                    />
+                                </Col>
+                            </Row>
+                        </>
+                    )}
+                </Grid>
+            </>
         );
     }
 }
