+++ src/popup/views/popupMap.js
@@ -5,8 +5,10 @@ import {
 import uuid from 'uuid';
 
 const mapStyles = {
-    width: '95%',
-    height: '95%',
+    width: '100%',
+    height: '400px',
+    'box-shadow': '0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19)',
+    'border-radius': '3px',
 };
 
 function arraysEqual(a, b) {
